Document and move SqliteSparseSequence
