+++ tests/common/sequences_test.py
@@ -1,9 +1,13 @@
+import os
+from tempfile import TemporaryDirectory
+
 import pytest
 
 from tango.common.sequences import (
     ConcatenatedSequence,
     ShuffledSequence,
     SlicedSequence,
+    SqliteSparseSequence,
 )
 
 
@@ -75,3 +79,24 @@ def test_concatenated_sequence():
     # __contains__()
     for item in l1 + [999]:
         assert_equal_including_exceptions(lambda: item in l1, lambda: item in l2)
+
+
+def test_sqlite_sparse_sequence():
+    with TemporaryDirectory(prefix="test_sparse_sequence-") as temp_dir:
+        s = SqliteSparseSequence(os.path.join(temp_dir, "test.sqlite"))
+        assert len(s) == 0
+        s.extend([])
+        assert len(s) == 0
+        s.append("one")
+        assert len(s) == 1
+        s.extend(["two", "three"])
+        s.insert(1, "two")
+        assert s[1] == "two"
+        assert s.count("two") == 2
+        ss = s[1:3]
+        assert list(ss) == ["two", "two"]
+        del s[1:3]
+        assert len(s) == 2
+        assert s[-1] == "three"
+        s.clear()
+        assert len(s) == 0
